modified:   src/libgcm/util/AnisotropicMatrix3DAnalytical.cpp 	modified:   src/libgcm/util/ThirdDegreePolynomial.cpp 	modified:   src/libgcm/util/ThirdDegreePolynomial.h 	modified:   src/tests/unit/anisotropic_matrixes_analytical.cpp

+++ src/libgcm/util/ThirdDegreePolynomial.h
@@ -8,11 +8,11 @@ namespace gcm {
 	{
 		/*
 		 * P(x) = x^3 + a*x^2 + b*x + c
-		 * All the roots are different 
-		 * or two ones are equal
-		 * " >= 0 " for all the roots
+		 * All the roots are real, different 
+		 * (or two ones are equal),
+		 * " > 0 " for all the roots
 		 */
-		float a, b, c;
+		double a, b, c;
 		float roots[3];
 		void findRoots();
 	public:
